Extract openssl RSA functionality into its own file.
